Main Site Update + Co-op Work Term Reports Addition
Added a headshot and navigation bar to the homepage, and removed the
favicon.

Co-op work term reports have been added and can now be accessed via the
navbar.

The S16 report has been added in full, and a stub page has been added
for the F16 report.

The resume has been temporarily removed as it has not yet been updated.
It will be added again in the next commit (of which the "Resume" button
on the navbar will link to).